* commit 'edc3fea2d374473527cdca7afd3229e4381fb969':
  VMLSINPOW-853 - Remove surcharge if any checkbox is selected
  VMLSINPOW-853 - Popup when min order not met and has checkbox selected

+++ SPApp/app/src/main/java/com/spgroup/spapp/domain/model/OrderSummary.kt
@@ -4,7 +4,7 @@ data class OrderSummary(
         val timestamp: Long,
         val partnerId: String,
         val orders: List<Order>,
-        val orderDiscountPercentage: Int,
+        val orderDiscountPercentage: Float,
         val overallCostBeforeDiscount: Float,
         val overallCostAfterDiscount: Float,
         val contactInfo: ContactInfo,
